Moved item sheet tabs into the runtime directory.
