+++ src/sheets/item/WeaponSheet.svelte
@@ -10,7 +10,7 @@
   import TextInput from 'src/components/inputs/TextInput.svelte';
   import Select from 'src/components/inputs/Select.svelte';
   import SelectOptions from 'src/components/inputs/SelectOptions.svelte';
-  import itemSheetTabs from '../itemSheetTabs';
+  import itemSheetTabs from '../../runtime/itemSheetTabs';
   import Source from '../shared/Source.svelte';
 
   let context = getContext<Readable<ItemSheetContext>>('context');
